+++ CHANGELOG.md
@@ -21,6 +21,7 @@ v0.6.0
  - `Binomial` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
  - `Bernoulli` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
  - `Geometric` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
+ - `Hypergeometric` now implements `Discrete<u64, f64>` rather than `Discrete<i64, f64>`
 
 v0.5.1
  - Fixed critical bug in `normal::sample_unchecked` where it was returning `NaN`